requirements.txt
numpy
opencv-python-headless
pandas
pillow
pytesseract
python-pptx
scikit-image
streamlit